remove paths that send team not found sentry error
