--- src/server/graphql/subscriptions/invitationSubscription.js
@@ -1,28 +1,0 @@
-import {GraphQLID, GraphQLNonNull} from 'graphql'
-import makeSubscribeIter from 'server/graphql/makeSubscribeIter'
-import InvitationSubscriptionPayload from 'server/graphql/types/InvitationSubscriptionPayload'
-import {getUserId, isTeamMember} from 'server/utils/authorization'
-import {INVITATION} from 'universal/utils/constants'
-import standardError from 'server/utils/standardError'
-
-export default {
-  type: new GraphQLNonNull(InvitationSubscriptionPayload),
-  args: {
-    teamId: {
-      type: new GraphQLNonNull(GraphQLID)
-    }
-  },
-  subscribe: (source, {teamId}, {authToken, dataLoader, socketId}) => {
-    // AUTH
-    if (!isTeamMember(authToken, teamId)) {
-      const viewerId = getUserId(authToken)
-      return standardError(new Error('Team not found'), {userId: viewerId})
-    }
-
-    // RESOLUTION
-    const channelName = `${INVITATION}.${teamId}`
-    const filterFn = (value) => value.mutatorId !== socketId
-    const resolve = ({data}) => ({invitationSubscription: data})
-    return makeSubscribeIter(channelName, {filterFn, dataLoader, resolve})
-  }
-}
